ref #107, Several improvements  (Thanks @sebcaux) Support for sticky-selection in replaceSelection methods. (Required for InpuMethod entry) Improved TextEditorComponent::InputMethodEvent... It now support special chars entry like expected. (Option+e,  e => ´ => é) - Fixed gapvector destructor: it did not use an array delete. - TextEditorWidget::setHorizontalScrollBar not emits the correct horizontalScrollBarChanged event.

+++ edbee-lib/edbee/views/components/texteditorrenderer.cpp
@@ -124,6 +124,8 @@ void TextEditorRenderer::renderLineBorderedRanges(QPainter *painter,int line)
     int lineHeight = renderer()->lineHeight();
 
     QPen pen(themeRef_->foregroundColor(), 0.5);
+//    QPen pen(themeRef_->findHighlightForegroundColor(), 0.5);
+//    QBrush brush(themeRef_->findHighlightBackgroundColor());
     painter->setRenderHint(QPainter::Antialiasing);
 
     int firstRangeIdx=0;
@@ -143,15 +145,15 @@ void TextEditorRenderer::renderLineBorderedRanges(QPainter *painter,int line)
             int startColumn = doc->columnFromOffsetAndLine( range.min(), line );
             int endColumn   = doc->columnFromOffsetAndLine( range.max(), line );
 
-            int startX = textLine.cursorToX( startColumn );
-            int endX   = textLine.cursorToX( endColumn );
+            qreal startX = textLine.cursorToX( startColumn );
+            qreal endX   = textLine.cursorToX( endColumn );
 
             if( range.length() > 0 && endColumn+1 >= lastLineColumn) endX += 3;
 
             QPainterPath path;
             path.addRoundedRect(startX, line*lineHeight + rect.top(), endX - startX, rect.height(),5,5);
+//            painter->fillPath(path, brush);
             painter->strokePath(path, pen);
-//            painter->strok(startX, line*lineHeight + rect.top(), endX - startX, rect.height(),  themeRef_->selectionColor() ); //QColor::fromRgb(0xDD, 0x88, 0xEE) );
         }
     }
 //PROF_END
